// rh-app/src/types/FormTypes.ts
export interface FormData {
  nome: string;
  email: string;
  telefone: string;
  data_nascimento: string;
  cpf: string;
  endereco: string;
  cidade: string;
  estado: string;
  cargo: string;
  experiencia: string;
  resumo_experiencia: string;
  disc: string[];
  ac1: string;
  ac2: string;
  comentariosFinais: string;
}